[ceremonyBox] move ceremony box sub-components to component directory
