+++ lib/page-encointer/ceremony_box/components/ceremonyNotification.dart
