+++ lib/page-encointer/ceremony_box/components/ceremonyProgressBar.dart
