+++ lib/page-encointer/ceremony_box/components/ceremonyRegisterButton.dart
